Move serialization code into sandbox
